+++ Runtime/Framework/State/IState.cs
@@ -17,63 +17,47 @@ namespace DT.UniStart {
     int BinarySearch(int index, int count, T item, IComparer<T> comparer);
   }
   public interface IDictionaryState<K, V> : IReadOnlyDictionary<K, V>, IWatchable, IWatchable<ReadOnlyDictionary<K, V>>, IGetValue<ReadOnlyDictionary<K, V>> { }
-
-  public interface ICommittableList<T> {
-    void Commit(UnityAction<IList<T>> action);
-  }
-  public interface ICommittableDictionary<K, V> {
-    void Commit(UnityAction<IDictionary<K, V>> action);
-  }
-
-  public interface IWritableState<T> : IState<T>, ISetValue<T> { }
-  public interface IWritableListState<T> : IListState<T>, IList<T>, ICommittableList<T> { }
-  public interface IWritableDictionaryState<K, V> : IDictionaryState<K, V>, IDictionary<K, V>, ICommittableDictionary<K, V> { }
   #endregion
 
-  #region State Manager
-  public interface IStateCommitter { }
-
-  public static class IStateCommitterExtension {
-    // Commit will trigger change event once.
-    public static IStateCommitter Commit<T>(this IStateCommitter self, IState<T> item, T value) {
-      (item as ISetValue<T>).Value = value;
-      return self;
-    }
-    public static IStateCommitter Commit<T>(this IStateCommitter self, IListState<T> item, UnityAction<IList<T>> f) {
-      (item as ICommittableList<T>).Commit(f);
-      return self;
-    }
-    public static IStateCommitter Commit<K, V>(this IStateCommitter self, IDictionaryState<K, V> item, UnityAction<IDictionary<K, V>> f) {
-      (item as ICommittableDictionary<K, V>).Commit(f);
-      return self;
-    }
-
-    // Apply may trigger change event multiple times.
-    public static IStateCommitter Apply<T>(this IStateCommitter self, IListState<T> item, UnityAction<IList<T>> f) {
-      f.Invoke(item as IList<T>);
-      return self;
-    }
-    public static IStateCommitter Apply<K, V>(this IStateCommitter self, IDictionaryState<K, V> item, UnityAction<IDictionary<K, V>> f) {
-      f.Invoke(item as IDictionary<K, V>);
-      return self;
-    }
-  }
-
-  public interface IStateManager : IStateCommitter { }
+  #region StateManager
+  public interface IStateManager { }
 
   public static class IStateManagerExtension {
-    // Add with interface checker
-    public static IState<T> SafeAdd<T>(this IStateManager _, IWritableState<T> state) => state;
-    public static IListState<T> SafeAddList<T>(this IStateManager _, IWritableListState<T> state) => state;
-    public static IDictionaryState<K, V> SafeAddDictionary<K, V>(this IStateManager _, IWritableDictionaryState<K, V> state) => state;
-
-    // helper methods, also check the interface of Watch family
-    public static IState<T> Add<T>(this IStateManager self, T value) => self.SafeAdd(new Watch<T>(value));
-    public static IListState<T> AddList<T>(this IStateManager self) => self.SafeAddList(new WatchList<T>());
-    public static IListState<T> AddList<T>(this IStateManager self, List<T> value) => self.SafeAddList(new WatchList<T>(value));
-    public static IListState<T> AddArray<T>(this IStateManager self, int n) => self.SafeAddList(new WatchArray<T>(n));
-    public static IListState<T> AddArray<T>(this IStateManager self, T[] value) => self.SafeAddList(new WatchArray<T>(value));
-    public static IDictionaryState<K, V> AddDictionary<K, V>(this IStateManager self) => self.SafeAddDictionary(new WatchDictionary<K, V>());
+    public static IState<T> Add<T>(this IStateManager manager, out Watch<T> echoed, T value) {
+      var state = new Watch<T>(value);
+      echoed = state;
+      return state;
+    }
+    public static IListState<T> AddArray<T>(this IStateManager manager, out WatchArray<T> echoed, int count) {
+      var state = new WatchArray<T>(count);
+      echoed = state;
+      return state;
+    }
+    public static IListState<T> AddArray<T>(this IStateManager manager, out WatchArray<T> echoed, T[] value) {
+      var state = new WatchArray<T>(value);
+      echoed = state;
+      return state;
+    }
+    public static IListState<T> AddList<T>(this IStateManager manager, out WatchList<T> echoed) {
+      var state = new WatchList<T>();
+      echoed = state;
+      return state;
+    }
+    public static IListState<T> AddList<T>(this IStateManager manager, out WatchList<T> echoed, List<T> value) {
+      var state = new WatchList<T>(value);
+      echoed = state;
+      return state;
+    }
+    public static IDictionaryState<K, V> AddDictionary<K, V>(this IStateManager manager, out WatchDictionary<K, V> echoed) {
+      var state = new WatchDictionary<K, V>();
+      echoed = state;
+      return state;
+    }
+    public static IDictionaryState<K, V> AddDictionary<K, V>(this IStateManager manager, out WatchDictionary<K, V> echoed, Dictionary<K, V> value) {
+      var state = new WatchDictionary<K, V>(value);
+      echoed = state;
+      return state;
+    }
   }
   #endregion
 }